Improve filters matching and reduce unnecessary calls
