+++ .changeset/late-plums-lick.md
@@ -1,0 +1,5 @@
+---
+"@interactors/core": patch
+---
+
+Improve filter matching and reduce unnecessary calls
